+++ src/main/java/org/mcdealer/mcdealer/Utils/DataManager/ConfigUpdater.java
@@ -1,7 +1,9 @@
 package org.mcdealer.mcdealer.Utils.DataManager;
 
+import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
 
+import org.bukkit.plugin.Plugin;
 import org.mcdealer.mcdealer.MCDealer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,10 +15,12 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 
 public class ConfigUpdater {
 
-    private static final Logger logger = LoggerFactory.getLogger("MCDealer (ConfigUpdater)");
+    private static final Logger logger = LoggerFactory.getLogger("MCDealer");
 
     private final MCDealer plugin;
 
@@ -28,9 +32,14 @@ public class ConfigUpdater {
         // Load the configurations
         JSONObject jsonConfig = loadJsonConfig();
         FileConfiguration yamlConfig = getConfig();
+        FileConfiguration villagerMarketConfig = getVillagerMarketConfig();
+
+        if (villagerMarketConfig == null) {
+            return;  // Exit if VillagerMarket plugin not found or config not loaded
+        }
 
         // Synchronize the values of the configurations
-        syncValues(yamlConfig, jsonConfig);
+        syncValues(yamlConfig, jsonConfig, villagerMarketConfig);
 
         // Save the updated JSON configuration
         saveJsonConfig(jsonConfig);
@@ -47,10 +56,23 @@ public class ConfigUpdater {
         }
     }
 
-    private void syncValues(FileConfiguration yamlConfig, JSONObject jsonObj) {
-        // Synchronize the values
-        jsonObj.put("currencySymbol", yamlConfig.getString("currencySymbol"));
-        jsonObj.put("currencySymbolPosition", yamlConfig.getString("currencySymbolPosition"));
+    private void syncValues(FileConfiguration yamlConfig, JSONObject jsonObj, FileConfiguration villagerMarketConfig) {
+        // Synchronize the values from VillagerMarket plugin config
+        String currencySymbol = villagerMarketConfig.getString("currency");
+        String currencySymbolPosition = villagerMarketConfig.getString("currency_before");
+
+        // Map the values from VillagerMarket config to the expected format
+        if (currencySymbolPosition != null) {
+            if (currencySymbolPosition.equalsIgnoreCase("true")) {
+                currencySymbolPosition = "before";
+            } else if (currencySymbolPosition.equalsIgnoreCase("false")) {
+                currencySymbolPosition = "after";
+            }
+        }
+
+        // Set the values in the JSON config
+        jsonObj.put("currencySymbol", currencySymbol);
+        jsonObj.put("currencySymbolPosition", currencySymbolPosition);
         jsonObj.put("defaultLanguage", yamlConfig.getString("defaultLanguage"));
     }
 
@@ -66,5 +88,37 @@ public class ConfigUpdater {
 
     private FileConfiguration getConfig() {
         return plugin.getConfig();
+    }
+
+    private FileConfiguration getVillagerMarketConfig() {
+        Plugin villagerMarketPlugin = Bukkit.getPluginManager().getPlugin("VillagerMarket");
+
+        if (villagerMarketPlugin != null) {
+            return villagerMarketPlugin.getConfig();
+        } else {
+            logger.warn("VillagerMarket plugin not found");
+            return null;
+        }
+    }
+
+    public static void pluginconfigupdater(FileConfiguration oldConfig, FileConfiguration newConfig) {
+        Map<String, Object> updates = new HashMap<>();
+        for (String key : newConfig.getKeys(true)) {
+
+            if (!oldConfig.contains(key)) {
+
+                updates.put(key, newConfig.get(key));
+            }
+        }
+        for (String key : oldConfig.getKeys(true)) {
+
+            if (!newConfig.contains(key)) {
+
+                updates.put(key, null);
+            }
+        }
+        for (Map.Entry<String, Object> entry : updates.entrySet()) {
+            oldConfig.set(entry.getKey(), entry.getValue());
+        }
     }
 }