+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.5] - 2024-09-24
+
+### Fixed
+- Inconsistent config variable (port)
+- Replaced ExternalIPFetcher with a more stable function 
+- Logback reconfigured
+
+### Added
+- Extended link command (Custom link in config)
+- Config Updater
+- Version Check
+
+### Changed
+- Updated dependencies
+- Updated Web files
+- Updated Jython script
+- Currency Symbol and Position will now be sync with Villagermarket config
+
 ## [1.1.0] - 2023-12-26
 
 ### Added